Dont even know

+++ notes/models.py
@@ -2,33 +2,23 @@ from django.db import models
 from django.contrib.auth.models import User
 
 
+class NotesFolder(models.Model):
+    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    title = models.CharField(max_length=100, default="Title")
+    recently_deleted = models.BooleanField(default=False)
+
+
 class Notes(models.Model):
     data = models.TextField()
-    user = models.ForeignKey(User, default=1, on_delete=set([1, ]))
+    user = models.ForeignKey(User, on_delete=models.CASCADE)
     title = models.CharField(max_length=50, default="title")
     added_time = models.DateTimeField(auto_now_add=True)
     last_edit_time = models.DateTimeField(default=None, null=True)
+    folder = models.ForeignKey(NotesFolder, on_delete=models.CASCADE, default=None, null=True)
+    deleted_at = models.DateTimeField(auto_now_add=False, null=True, default=None)
 
     @staticmethod
-    def get_notes_by_ranged_title(user, title_range=list()):
-        """
-        @brief
-        This function get notes by ranged title
-        """
-        notes = Notes.objects.filter(user=user)
-        sorted_list = []
-        if len(title_range) == 1:
-            for i in notes:
-                if i.title[0].lower() == title_range[0].lower():
-                    sorted_list.append(i)
-        else:
-            for i in notes:
-                if title_range[0].lower() <= i.title[0].lower() <= title_range[1].lower():
-                    sorted_list.append(i)
-        return sorted_list
-
-    @staticmethod
-    def get_notes(sorting_type, user=1, notes=None):
+    def get_notes(sorting_type, user=None, notes=None, folder=None):
         """
             @brief
             This function get notes
@@ -38,6 +28,8 @@ class Notes(models.Model):
             if aim = 'title' -> 'up' = a-z, 'down' = z-a
             if sorting type is not correct returns returns notes sorted like date_up
         """
+        if user is None:
+            return None
         if notes is None:
             notes = Notes.objects.filter(user=user)
         notes = notes.filter(user=user).order_by('-added_time')
@@ -66,17 +58,16 @@ class Notes(models.Model):
                 return notes
         else:
             return notes
+        recently_deleted_folder = NotesFolder.objects.filter(user=user, recently_deleted=True)
+        if len(recently_deleted_folder) == 0:
+            tmp_folder = NotesFolder(user=user, title="Recently deleted", recently_deleted=True)
+            tmp_folder.save()
+        recently_deleted_folder = NotesFolder.objects.filter(user=user, recently_deleted=True)
+        if folder is None:
+            notes = notes.exclude(folder=recently_deleted_folder[0])
+        else:
+            notes = notes.filter(folder=folder)
         return notes
-
-    @staticmethod
-    def get_note_by_id(note_id):
-        """
-        @param
-        This is ID of notes
-        @brief
-        This function gets note by id
-        """
-        return Notes.objects.filter(id=note_id).first()
 
     @staticmethod
     def delete_note(note_id, user):
@@ -88,38 +79,31 @@ class Notes(models.Model):
         """
         notes = Notes.objects.filter(user=user, id=note_id)
         if len(notes) > 0:
-            notes[0].delete()
+            fully_deleted = False
+            if notes[0].folder:
+                if notes[0].folder.recently_deleted:
+                    notes[0].delete()
+                    fully_deleted = True
+            if not fully_deleted:
+                recently_deleted_folder = NotesFolder.objects.filter(user=user, recently_deleted=True)
+                if len(recently_deleted_folder) == 0:
+                    tmp_folder = NotesFolder(user=user, title="Recently deleted", recently_deleted=True)
+                    tmp_folder.save()
+                recently_deleted_folder = NotesFolder.objects.filter(user=user, recently_deleted=True)
+                notes[0].folder = recently_deleted_folder[0]
+                notes[0].save()
             return 100
         else:
             return 111
 
     @staticmethod
-    def search_notes(string, user, sorting_type="date_up"):
+    def search_notes(aim, user, sorting_type="date_up", folder=None):
         """
         @brief
         This function searches for the notes
         """
-        obj = Notes.get_notes(sorting_type, user)
-        ret_list = list()
-        for i in obj:
-            if string in i.title:
-                ret_list.append(i)
-        return ret_list
-
-    @staticmethod
-    def notes_sort_by_date(datetime, user):  # note: datetime = {1: date_one NOT NULL, 2: date_two}
-        """
-        @param
-        This is the time at which the notes are sorted
-        @brief
-        This function sorts notes
-        @detailed
-        This function sorts notes by date and alphabetically and aslo from new to old and conversely
-        """
-        notelist = Notes.objects.filter(user=user)
-        if len(datetime) == 1:
-            date = datetime[0].date()
-            return notelist.filter(pub_date=date)
-        else:
-            return notelist.filter(pub_date__gte=datetime[0].date(),
-                                   pub_date__lte=datetime[1].date()).order_by('-pub_date')
+        notes = Notes.get_notes(sorting_type, user, folder=folder)
+        if aim is None:
+            aim = ""
+        notes = notes.filter(title__icontains=aim)
+        return notes
